fix: added policies to staticData endpoints + fix linter
