+++ sources/core/Ggio.BikeSherpa.Backend.Domain/DeliveryAggregate/Bootstrap.cs
@@ -1,5 +1,3 @@
-using Ggio.BikeSherpa.Backend.Domain.DeliveryAggregate.Enumerations;
-using Ggio.BikeSherpa.Backend.Domain.DeliveryAggregate.PricingStrategies;
 using Ggio.BikeSherpa.Backend.Domain.DeliveryAggregate.Services.PricingStrategy;
 using Microsoft.Extensions.DependencyInjection;
