--- sources/core/Ggio.BikeSherpa.Backend/Features/Public/PackingSizes/Model/PackingSizeDto.cs
@@ -1,7 +1,0 @@
-namespace Ggio.BikeSherpa.Backend.Features.Public.PackingSizes.Model;
-
-public record PackingSizeDto
-{
-     public required string Label { get; set; }
-     public required string Value { get; set; }
-}
