--- sources/core/Ggio.BikeSherpa.Backend/Features/Public/PricingStrategies/GetAll/GetAllPricingStrategiesHandler.cs
@@ -1,34 +1,0 @@
-using Ggio.BikeSherpa.Backend.Domain.DeliveryAggregate.Enumerations;
-using Ggio.BikeSherpa.Backend.Features.Public.PricingStrategies.Model;
-using Mediator;
-
-namespace Ggio.BikeSherpa.Backend.Features.Public.PricingStrategies.GetAll;
-
-public record GetAllPricingStrategiesQuery() : IQuery<List<PricingStrategyDto>>;
-
-public class GetAllPricingStrategiesHandler : IQueryHandler<GetAllPricingStrategiesQuery, List<PricingStrategyDto>>
-{
-     public async ValueTask<List<PricingStrategyDto>> Handle(GetAllPricingStrategiesQuery query, CancellationToken cancellationToken)
-     {
-          var strategies = new List<PricingStrategyDto>
-          {
-               new()
-               {
-                    Label = "Personnalisé",
-                    Value = PricingStrategy.CustomStrategy
-               },
-               new()
-               {
-                    Label = "Course",
-                    Value = PricingStrategy.SimpleDeliveryStrategy
-               },
-               new()
-               {
-                    Label = "Tournée",
-                    Value = PricingStrategy.TourDeliveryStrategy
-               }
-          };
-
-          return strategies;
-     }
-}
