--- sources/core/Ggio.BikeSherpa.Backend/Features/Public/Urgencies/Model/UrgencyDto.cs
@@ -1,7 +1,0 @@
-namespace Ggio.BikeSherpa.Backend.Features.Public.Urgencies.Model;
-
-public record UrgencyDto
-{
-     public required string Label { get; set; }
-     public required string Value { get; set; }
-}
