+++ sources/core/Ggio.BikeSherpa.Backend/Features/StaticData/PricingStrategies/Model/PricingStrategyDto.cs
@@ -1,0 +1,11 @@
+using Ggio.BikeSherpa.Backend.Domain.DeliveryAggregate.Enumerations;
+using JetBrains.Annotations;
+
+namespace Ggio.BikeSherpa.Backend.Features.StaticData.PricingStrategies.Model;
+
+public record PricingStrategyDto
+{
+     [UsedImplicitly]
+     public required string Label { get; set; }
+     public required PricingStrategy Value { get; init; }
+};
