+++ sources/core/Ggio.BikeSherpa.Backend/Features/StaticData/Urgencies/Model/UrgencyDto.cs
@@ -1,0 +1,7 @@
+namespace Ggio.BikeSherpa.Backend.Features.StaticData.Urgencies.Model;
+
+public record UrgencyDto
+{
+     public required string Label { get; init; }
+     public required string Value { get; init; }
+}
